[FIX] sale_renting_planning: avoid breaking loop
Before this commit, when rental order containing a plannable rental
product is confirmed, some shifts will be generated according to the
rental order lines with such product. The problem is when there is no
available resource, the method in which we expect to return vals list of
planning.slot model will return vals because it will break the loop
instead of continuing to the next iteration.

This commit makes sure the `_planning_slot_vals_list` always return list
of vals.

closes odoo/enterprise#92058

X-original-commit: fa2583bd777bf7ff9f3972fe3c174b44ca4519cf
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>