[FIX] l10n_es_reports: change date format to dd/MM/yyyy
Changed date format from MM/dd/yyyy to dd/MM/yyyy in reports.

An issue with the date formatting for the exported Excel file of the Generic Tax Return report for Spain.
In the XLSX file, the dates are generated in the mm/dd/yyyy but they should be generated in the dd/mm/yyyy format.

Steps to reproduce this error:

Create a new database using Odoo version 18.
Activate the module: “Spain - Accounting (PGCE 2008)” (l10n_es).
Create a new company and set Spain as the country.
Create a new quotation and select the company created in step 3.
Confirm the invoice generated from the quotation.
Go to Accounting → Reporting → Tax Report and select "Generic Tax Report".
Click the gear icon and select "VAT Record Books (XLSX)".

closes odoo/enterprise#93458

X-original-commit: f6b484e1d72f6f74cb27ab1632aa1e849c79b444
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>