[FIX] partner_commission: make commission plan read-only in upsell orders
Problem:
An inconsistent behavior occurs when a user changes the commission
plan while creating an upsell order for a recurring Sales Order (SO).

Steps to reproduce:
1) Install the Subscriptions and partner_commission modules.
2) Create a subscription SO with a referrer_id and set a recurrence.
3) Enable the Freeze Plan option and create a commission plan from the view
4) Add a rate of 50 and product category as service in the commission rules.
5) In SOL add a subscription product contains recurrence with unit price of 100
6) Confirm the SO → Create Invoice → Confirm → Pay.
7) Go back to the SO and create an upsell for it.
8) Change the commission plan rate to 30 by creating a new plan.
9)Repeat step 4.

Issue:
When you navigate to the referrer record from the SO and open the
Purchase Order via the smart button, you will see two Purchase Order lines
both showing a value of 50, even though the upsell order had a new commission
plan with a rate of 30.

Cause:
When the invoice is marked as paid, a Purchase Order with POL is created
using values from the commission plan.

For subscription orders, the system intentionally uses the subscription’s
original commission plan instead of the updated one.

However, the problem is that the commission plan field remains editable in
upsell SOs even when the parent SO has Freeze Plan enabled, misleading users
into thinking the new commission plan will be applied.

Solution:
Make the commission plan field read-only for upsell SOs when the parent SO
has Freeze Plan enabled.

opw-4954307

closes odoo/enterprise#92723

X-original-commit: baf1fac0d16c04e19569564a36d45d597fb3ccc6
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>