[IMP] web_cohort: improve breadcrumb restoration
Since [1] when reloading the browser the breadcrumb is restored based on the
URL.

This was implemented in a very naive way. For each action found in the URL, a
multiview was added to the breadcrumb. Similarly, for each pain action/ID found
in the URL, a form view was added to the breadcrumb. Note that the active_id
(found before the action in the URL), is taken into account in the context of
the action.

For example, if the URL is : `/odoo/project/5/tasks`:
 - for the `project` action, the list of projects will be added to the
    breadcrumb;
 - for the `project/5` action/ID pair, the form view of the project 5 will be
 added;
- for the `5/tasks` action (with active_id 5), the list of tasks for the
project 5 will be added.

However, this naive implementation does not consider that the user can navigate
directly from the list of projects to the list of tasks for a project without
passing through the project form view.

This commit saves the current state in the session storage, to be used to
restore the breadcrumb when the page is reloaded.

Note that similar behaviour was implemented to improve action restoration [2].

[1] https://github.com/odoo/odoo/commit/c63d14a0485a553b74a8457aee158384e9ae6d3f
[2] https://github.com/odoo/odoo/commit/e8afebdcbf5d725091983fbe3b425fec4aa3a071

task-id 5011505

closes odoo/enterprise#92168

Related: odoo/odoo#222605
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>