[CLN] hr_expense_predict_product: remove OCR category prediction code
In the newer version of the expense OCR (see commit 7c966fc), it doesn't
predict a category anymore. Instead, it actually detects a description
from the expense that can be used to predict the product.

This commit cleans up the code that relates to the category prediction
as it's not used anymore.

closes odoo/enterprise#92123

Related: odoo/upgrade#8208
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>