[FIX] l10n_de_reports: missing version strings in EC Sales list csv
When uploading the EC Sales List csv to Elster.de portal, users will see
the file rejected with the following error (EN translated)

"""
Your CSV file was created for an older version of the CSV interface.
Please refer to the version table in the import help for the latest
version of the interface and any changes. If the version information
is correct, please check the encoding of your CSV file. The permitted
encoding is UTF-8.
"""

This occurs because the csv is missing the version strings identifying
the versions of data format

References:
https://www.elster.de/eportal/helpGlobal?themaGlobal=zmdo_import_eop
https://www.datev-community.de/t5/Betriebliches-Rechnungswesen/Import-CSV-Daten-nacvh-Elster-f%C3%BCr-ZM/td-p/480079

opw-4719033

closes odoo/enterprise#92344

X-original-commit: 4c3abf0eb21703541934560170a565d8d84921cc
Signed-off-by: William André (wan) <wan@odoo.com>