[IMP] accountant_knowledge: add title page for empty articles
Currently, the system skips printing empty articles, which can lead to
an incomplete table of contents and a less organized PDF structure. To
address this, we will now generate a page with the article's title for
empty articles. This approach allows empty articles to serve as section
dividers, enhancing the document's organization and improving navigation
within the PDF.

Task-4989809

closes odoo/enterprise#92120

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>