[FIX] ai_documents: prevent ai sort prompt issues when edited too quickly
Purpose:
--------
When the AI sort prompt is edited too quickly, the focus inside the prompt
widget is lost, and some characters are not inserted.
This happens because the record is updated on every char change so that
the show_incomplete_prompt_warning field can be recomputed to show/hide
the warning.
When the prompt is edited too quickly, onWillUpdateProps is then called
with a prompt value that does not match the current editor value, leading
to a manual editor update with the old value.

Now, the condition to show/hide the message is computed in frontend, and
the record is only updated when the editor is blurred.

Task-4989962

Part-of: odoo/enterprise#91842
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>