[IMP] appointment: enable min schedule hours for custom appointments
Before, we considered custom appointments to be managed precisely
and therefore did not account for min_schedule_hours. Even, we
let invited customers book a slot that was already started, the
only restriction was that it was not ended.

As custom appointments are easier to configure and will probably
be more frequent and used in various contexts, now that we have
added a 'flexible' schedule option directly on the form, we make
sure that the min_schedule_hours is working as well, in a standard
way, for custom slots. Add a test case as well.

Task-4997980

Part-of: odoo/enterprise#91824
Related: odoo/upgrade#8206
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>