[FIX] l10n_mx_edi_pos: Use currency rounding to round amount total
Issue: With specific prices and discounts, we can get an rounding issue
that does not pass the Validation Error.

Purpose of this PR:
To round the amount total based on the currency's rounding precision to pass the check.

Steps to Reproduce on Runbot:
install pos, l10n_mx
create products prices with 75.00 and 60.00 respectively.
create discount 100% on order
change decimal accuracy to 3, change tax to Include with Price

create PoS order with products and apply discount
receive ValidationError even though the amount total should be rounded to 0.0.

opw-4944830

closes odoo/enterprise#92318

X-original-commit: 80d13fecfbc5971f8f50aeef4a29408f11a78a72
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>