[FIX] account_accountant: batch payment negative statement line
When having a negative statement line that we want to reconcile with a batch
payment composed with payments from bills. The different line of the batch
payment were not correctly split and so the bills were not put as paid.

This was because of the use of the min that in case of negative amount was
taking the amount of the entire batch and not the amount of the payment itself.

task-4749334

Part-of: odoo/enterprise#88327
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>