[FIX] account_accountant, account_accountant_batch_payment: batch payment button
When selecting or deleting a batch payment, the button was not updated
correctly. For example, having just one batch and selecting it should make the
button disappear for every statement line.
Also, the amount of the batch was wrong since it was not the residual amount.

To fix this, we move the logic of to get the different batch payment to the
service and use that to update the available batch payment when selecting one or
deleting one.

task-4749334

closes odoo/enterprise#88327

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>