[FIX] documents_project: update tooltip to description current behavior
Before this commit, the tooltip displayed for `documents_folder_id`
defined in `project.project` model, said all documents linked to a task
of a project will be automatically added inside that folder but it is
no longer the case since the refactoring of Documents to simplify the
implementation but also let the documents user sorts his documents as he
wants since now the access rights are defined by folder.

This commit updates the tooltip accordingly.

PR simplifying the bridge between documents and project apps: #75214

opw-5000224

closes odoo/enterprise#92501

X-original-commit: 7f38061a8cdfe962dc4c1c4859a1a5a6013f057e
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>