[FIX] accountant_knowledge: add pdf extension to output report
Currently, when a user exports an audit report, the server returns a
download URL. During the download, the server sets the `Content-Type`
HTTP header to `application/pdf`, indicating that the file is a PDF.

However, for Windows users, setting the `Content-Type` header alone is
not enough. The operating system may treat the file as a generic file,
which means it won't automatically open in a PDF viewer when double-clicked.
Instead, the user must manually assign the `.pdf` extension or choose a
PDF reader to open it.

To resolve this, we will include the `.pdf` extension in the file name.
This ensures that the operating system correctly identifies the file as
a PDF and opens it with the appropriate application.

Task-4989809

closes odoo/enterprise#92166

Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>