[FIX] mrp_plm: count ECOs waiting for me correctly in dashboard
The dashboard of PLM displays for each ECO type the number of
validations pending in general, and the validations waiting for me
specifically.

The latter has been incorrectly computed for quite some time, as it
computes how many ECOs exists where I am an approver and an approval is
waiting - but there is no guarantee that the waiting approval is one of
mine; it might be that I am a past approver of the ECO and that another
approval (waiting for somebody else than me) exists, or that two
approval are required on the same stage (one by me, one by somebody
else) and I already approved.

This commit switches the computation around by looking for pending
approvals waiting for me, and then get the number of (non-archived) ECOs
bound to them.

closes odoo/enterprise#93866

X-original-commit: fa3ca45c8f101cb06115a165d9d7b12312b252e1
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Bouvy Damien (dbo) <dbo@odoo.com>