[IMP] *: Automatic transfers revamp
*: account_auto_transfer, account_transfer

Description of the issue this commit addresses:

The automatic transfers have been the way they are for a long time now and are
barely used compared to how they could be. A little revamp could give them a
bigger impact and use.

Desired behavior after this commit is merged:

The automatic transfers have a renewed UI and use different systems to redirect
the amounts of the original moves.

upgrade pr: https://github.com/odoo/upgrade/pull/7887
task-4813376

closes odoo/enterprise#87015

Related: odoo/upgrade#7887
Signed-off-by: John Laterre (jol) <jol@odoo.com>