[FIX] account_payment: correct chatter log when validating batch payment
When validating a batch that includes payments in process (e.g., without journal entries),
Odoo incorrectly logs that the payment was removed from the batch, even though it remains.

This commit fixes the chatter message by not possessing the write method of the account_payment if batch_payment_id not in vals.

task-5011672

closes odoo/enterprise#92596

X-original-commit: bd949d9a93e4ad630cecff6ddb5fbe6704ce1901
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Ahmed Elemary (aele) <aele@odoo.com>