[FIX] l10n_ar_edi: don't keep retrying test setup forever
In 18.0 and above test hasn't worked in months, half the time it fails
because pretty much nothing works in it and half the time it's because
it keeps looping on trying to set up the connection and
failing (possibly because we're getting blacklisted due to how much
spamming we do).

For now restrict to just trying 5 times (per class since it's called
from a `setUpClass`) with an exponential backoff. This may yet end up
in removing the entire test directory as it doesn't seem like anyone
cares.

closes odoo/enterprise#92281

X-original-commit: c4f38f0a47819f173eda3cbdcbba31ac7a90ad43
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>