[FIX] documents_spreadsheet: ensure portal users cannot be 'edit' role
Let's say a Document Folder is shared with a portal user with 'edit'
access. The portal user shares a .xlsx file to the folder, and an
internal user later try to convert the file to odoo spreadsheet.
During the conversion, the portal user has 'edit' role on the folder,
which is copied to the documents.access records of the converted sheet.
Since Odoo prevents Spreadsheets from being shared in edit mode to
portal users, _check_spreadsheet() raises a Validation Error.

To resolve this issue, we force portal users to have `view` role on
Spreadsheets during the XLSX to Spreadsheet conversion.

opw-4753670

closes odoo/enterprise#92465

X-original-commit: 3b2bfb4dc93a0d79758458d60e1d083445bb3f06
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>