[FIX] sale_subscription_timesheet: get hours before next_invoice_date
Problem: When the user invoices a product based on timesheet from a
subscription in a given month, the timesheets linked to the invoice
incorrectly includes timesheets dated the same as the next invoice date
of the subscription. This occurs if the client does not specify a timesheet
period when creating the invoice. For example, the user invoices a
subscription for the period of July 1st to July 31st and the next invoice
date of the subscription is August 1st. There are timesheets dated in July and
a timesheet dated August 1st. The quantity delivered for the invoice will
only include timesheets within the July period. However, the linked
timesheet to the invoice will mistakenly include the timesheet from
August 1st.

Purpose: Although the linked timesheet has no effect on the computation
of qty_delivered for the current period, the timesheets dated the same as the
next_invoice_date will get linked to the invoice and no longer be considered
as delivered hours for the next period, which could cause gaps in hours
delivered. The next_invoice_date should not be
included in the date range when linking timesheets within the period of the
invoice, since these hours were not included in the computation of quantity
delivered.

Steps to reproduce on Runbot:
1. Install Sales, Project, Timesheet
2. Enable Timesheet in Project > Settings
3. Create a service product that is based on timesheet
and creates project and task on order
4. Create a quotation for the service product and
set the recurrence to be monthly and
set the start date to be the first of the past month
5. Confirm the quotation and
set the next invoice date to be the first of the current month
6. Create timesheet hours within the period and one dated
to be the first of the current month
7. Create an invoice for the subscription and
notice that the timesheet from the first of the current
month is linked to the invoice.

opw-4689894

closes odoo/enterprise#92445

X-original-commit: f308a422dc895539433ac3f8ea48ae97483e72d6
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Mylyna Hy (myhy) <myhy@odoo.com>