[FIX] timesheet_grid_holidays: freeze date in test
Before this commit, when the
`test_timer_methods_handle_project_access_restrictions` test is
executed after 22h in UTC the test failed because the timezone of the
current user is the belgian one.

This commit makes sure the date and timezone no longer affects this test
since it is not the purpose of this test.

runbot-error-230490

closes odoo/enterprise#92498

X-original-commit: b096d27a13a0c106e42f539edf947129466b3d7a
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>