[FIX] ai_fields: only assign empty strings to textual fields
Purpose:
--------
Currently, when an automation rule is added with "update with ai" on a
relational field and when the query in unresolvable, an error is raised
because an empty string is assigned to the field.

The empty string will now only be the assigned to textual fields (which
is required to not run the ai fields cron infinitely on an unresolvable
query).

Task-4989962

Part-of: odoo/enterprise#91842
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>