[FIX] l10n_it_pos: print past orders
before this commit, when trying to reprint a past order it would not work.
Instead it would print the last order no matter what order is selected.

Steps to reproduce:
1. setup an DB with an italian POS
2. make 2 sales with a different product (easier tracking)
3. open the "orders" view
4. try to reprint the ticket of the firt order
result: the receipt of the second order is printed

the reason for this is that we were using the printer's built in command
"printDuplicateReceipt" which is inteded for printing the very last receipt.

With this commit, we changed the behavior and invoke another printer command.
That command is meant to reprint any receipt, based on the provided reference.
After this commit, trying to print any past receipt will print that exact receipt.

IMPORTANT NOTE:
In theory, the printer command can reprint any number of receipts.
We decide that we will only use it to reprint the one receipt selected by the user.
This moves the complexity of parsing date strings and ranges to the command component and
therefore will keep the rest of the code cleaner.

opw-5008702
opw-4882480

closes odoo/enterprise#99946

X-original-commit: fcb0a7477a3b8668ce68af6ed4f18535127d6496
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>
Signed-off-by: Victorien Sonneville (viso) <viso@odoo.com>