[FIX] l10_ke_edi_edi_oscu: ensure quantity and monetay values checks are performed on reconciled reversals only.
Fixes a bahavior where the checks performed on the credit notes related to thier quantities and monetary values include non reconciled credit notes.

Steps to reproduce:

1 - activate `l10n_ke` on some company.
2 - Create an invoice.
3 - Create a partial or full credit note ( this one will be reconciled with the invoice automatically )
4 - cancel the credit note or reset it to draft.
5 - create another credit note where the quantities and/or monetay values exceed that of the invoice if summed up with the cancelled credit note.

Following the steps will result in an error message saying the the monetary value or quantities on the credit notes exceed that of the invoice.

The correct behavoir is to simply not count any credit note that isn't explicitly reconciled with the invoice.

opw-4779976

closes odoo/enterprise#93026

X-original-commit: 531e26b4e1cf7dbfe90e75df3bc979813a47946c
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>