[FIX] hr_expense_extract: fix expense name not being set by OCR
The combination of the following recent changes prevented the OCR from
filling the name of the expense:
- New expenses are named "Untitled Expense {date}" by default instead of
  the name of the attachment (see commit odoo/odoo@ca1f644).
- The OCR doesn't override the name if it's not set to the default value
  (see commit 5c813dc).

The first commit didn't update the default expense name that was used in
the OCR code, it was still computed from the attachment name.

After the second commit, the OCR couldn't fill in the name of the
expense anymore, since it was expecting the name to be the attachment
name, but it was "Untitled Expense {date}" instead.

This is fixed by checking that the expense name contains "Untitled
Expense" (and taking translation into consideration).

This commit also improves the tests in multiple ways:
- For the `test_auto_send_for_digitization` test, it now mimicks an
  upload of the document to force the creation of the expense with the
  default name, this should prevent this from happening again.
- It now uses an actual PDF document instead of "fake" attachments.
- Minor clean up.

task-4684825

closes odoo/enterprise#92585

X-original-commit: 667255089176dad3bbcc7b00c9da6de1d55d3983
Related: odoo/odoo#223303
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>