[FIX] account_iso20022: pass schme_nm on override
Sweden's implmentation of SEPA/ISO20022 overrides
`_get_company_PartyIdentification32` to add SchemeName support in the
node. The issue is that this override doesn't pass the `schme_nm`
parameter to the base implementation which performs the same action.
This commit fixes that by adding the parameter in the super and removing
the duplicated code as it now causes the node to appear twice.

task-4939347

Part-of: odoo/enterprise#90378
Related: odoo/odoo#219352
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>