[IMP] l10n_br_avatax: support overriding cnae per line
Some cities require specific CNAE codes to be set on services to
determine taxes, benefits, and taxpayer classification. Avalara does not
maintain a central table to automatically apply the right CNAE code per
city per service as such we must have a way of representing this on the
line level.

Rather than adding another field on `account.move.line` we can take use
of `l10n_br.ncm.code` and add a link to `l10n_br.cnae.code` within the
model. This allows products to have specific NCM codes by city and put
the right CNAE code on each line.

In addition, if new CNAE codes are created or needed by certain states
or cities that we don't currently have, a new menu item and view was
created to allow for creation or editing of the codes. As such a
uniqueness constraint was also added.

task-4950444

closes odoo/enterprise#90703

Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>