[FIX] website_generator: add the button to select custom footer template
With the initial [website builder refactor], the possibility to select
the custom footer template has been lost. This commit adds it again with
a resource. Instead of modifying the view with the list of templates
when one is created, it queries the server to know whether there is one

[website builder refactor]: 9fe45e2b7ddbbfd0445ffe25a859e67a316d02b2
task-4367641

closes odoo/enterprise#93602

Related: odoo/odoo#224821
Signed-off-by: Christopher du Toit (chto) <chto@odoo.com>