[FIX] l10n_mx_edi: Fix mismatch between invoice post time and CFDI fecha
'fecha_datetime' could be set to a different value than the
'document_post_time' passed as parameter because of the "min".

We also must revert the `_cr.commit` during the sending because it
removes the lock on records. Instead we set the invoice post time and
commit before locking.

task-none

closes odoo/enterprise#92938

X-original-commit: b943b8f44dac8de23e970d249ce822aa065aaeab
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>