[IMP] stock_barcode(_mrp); display all SN
Now we can produce multiple product tracked by SN with only one MO but
in the Barcode app, we display the MO's produced serial number only if
there is a single SN.
This commit changes that to display the MO's SN regardless the SN
amount. That said, the header line can contain a lot of serial numbers
(imagine a product for 20 units of a product tracked by SN), so to avoid
a huge ugly text block, if there is more than 5 serial numbers to
display, we display only the three first serial numbers and the last one
with an ellipsis between.

task-4830458

Part-of: odoo/enterprise#90930
Related: odoo/odoo#220383
Related: odoo/upgrade#8117
Signed-off-by: Steve Van Essche <svs@odoo.com>