[IMP] stock_barcode_mrp: generate SN button
Before this commit, the button to generate serial numbers for a MO in
the Barcode app was visible only if the MO has no SN. Now, it will be
visible as long as there is not enough generated serial numbers
regarding the quantity to produce.

Also, this commit set the button as secondary once at least 1 SN was
generated, since we assume if the user generated less than needed, maybe
they don't want to generate them anymore.

closes odoo/enterprise#90930

Related: odoo/odoo#220383
Related: odoo/upgrade#8117
Signed-off-by: Steve Van Essche <svs@odoo.com>