[FIX] spreadsheet_edition: allow to remove invalid list sorting
If a list is sorted on an invalid field and you try to open
the list side panel, it crashes.
It should open to allow the user to remove the sorting.

A list with an invalid sorting field if the spreadsheet was created
in a given version, then upgrades to a version where that field
has been removed or renamed.

closes odoo/enterprise#96069

Task: 4962837
X-original-commit: 1b8e1784f18a22a7b8009d18ea1e28d087e6f945
Related: odoo/odoo#229610
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>