[FIX] account_invoice_extract: select VAT matched partner by their rank
It's possible that multiple partners are sharing the same VAT number,
for example, when multiple addresses are set on it (each address will be
a child `res.partner` with the same VAT number).

Previously, when the OCR matched such partners from their VAT number,
it would select one "at random" (based on the default sort of
`res.partner`).
Now, it will match the one with the highest supplier/customer rank.

opw-4954057

closes odoo/enterprise#93153

X-original-commit: 37a981fff632ccd5c7a5ab7bdf08b3d26fd11c08
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>