[FIX] l10n_nl_reports: fix tax payment communication
Steps to reproduce:
- Install Accounting and l10n_nl
- Switch to a Dutch company (e.g. NL Company)
- In company form, set VAT and Omzetbelastingnummer

- Create an invoice with tax for previous month
- Confirm the invoice

- Go to "Accounting / Reporting / Statement Reports / Tax Return"
- Select previous month
- Create "Closing Entry" and post it

- On "Miscellaneous Operations" journal in Accounting dashboard, click
on "Pay tax: [date]" link
- A wizard will appear fot the VAT payment

Issue:
The value of the "Communication" field is incorrect.

Cause:
The "Communication" field is always computed from the VAT of the company.
However, when the "Omzetbelastingnummer" field is set, it should be used
instead of the VAT.

Solution:
Use the "Omzetbelastingnummer" field to compute the "Communication" of
the VAT payment when it is set.

opw-4836615

closes odoo/enterprise#92766

X-original-commit: d64ae299e2e348181399dbea202bd7cb856d497b
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>