[FIX] website_appointment: fix dynamic snippets drag and drop test
Starting from [1], the various dynamic snippet layouts are directly
accessible from the snippet selection dialog instead of using the
"Templates" option.

The goal of this commit is to simply add the new snippets keys in the
"sub-snippet templates" config to prevent the snippets drag and drop
test from failing.

[1]: https://github.com/odoo/odoo/commit/e3b062e5d3820ddfcee2eb669f21edc0c53c3330

runbot-230822

closes odoo/enterprise#92392

Related: odoo/odoo#223010
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>