[FIX] helpdesk_timesheet: fix check for default company_id
Steps to Reproduce:

- Open Helpdesk.
- Go to Configuration → Helpdesk Teams.
- Create a new team.
- Enable the Billing feature and assign a project to this team.
- Set the project’s Company field to Null.
- Create an employee for the current user in another company.
- Ensure the user has employees in both companies.
- Create a new ticket (or open an existing one) in the newly created team.
- Try creating a new timesheet → a Validation Error is raised.

Root Cause:
When a user has employees associated with multiple companies and tries to log a
timesheet with multi-company enabled, a validation error occurs.

This happens because:

- The default company is derived from the project linked to the
Helpdesk team. Since the company_id field on the project is no
longer required, it may be Null, leading to an error when fetching
the correct employee for the company.
https://github.com/odoo/odoo/blob/6a36015e2ee69aafe3880bf1fff38439af5cd673/addons/hr_timesheet/models/hr_timesheet.py#L214-L216

- Additionally, the company check is incorrect. The code is checking
whether the `company_id` key exists in a `list of vals`, whereas
it should be checked directly on the vals dict itself.
https://github.com/odoo/enterprise/blob/83a1b88c8cae4db5c9b1bf82ee7ebb4e13c41b2a/helpdesk_timesheet/models/analytic.py#L81-L82

Issue Faced:
In version 18.3, timesheets are created during migration here.
https://github.com/odoo/upgrade/blob/24f85bbc3408bf10b1cee93e4c395edf806beaa8/migrations/helpdesk_timesheet/saas~18.3.1.0/post-migrate.py#L44-L58
Even though the correct company_id is passed, it gets overridden
during the process. and If the associated project does not have a
company set, the company_id becomes False,
https://github.com/odoo/enterprise/blob/83a1b88c8cae4db5c9b1bf82ee7ebb4e13c41b2a/helpdesk_timesheet/models/analytic.py#L81-L82
which results in a Validation Error because the here unable to
determine the correct employee linked to the company.
https://github.com/odoo/odoo/blob/6a36015e2ee69aafe3880bf1fff38439af5cd673/addons/hr_timesheet/models/hr_timesheet.py#L214-L216

closes odoo/enterprise#95411

Opw: 5004092
X-original-commit: 1c5e547b59a57c840ef72b2e76c6a6a627f1f18e
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>