[FIX] l10n_mx_edi: synchronize fiscal regime PDF and XML
In l10n_mx:
- Create a child contact under a company contact.
- Set the fiscal regime of the child contact to one different from the company’s fiscal regime.
- Create an invoice with the child contact and send it to the CFDI.

In the XML, the fiscal regime used is the company’s, whereas in the PDF it is the child contact’s.
This commit applies the same logic from _add_customer_cfdi_values to the PDF generation.
After this change, the fiscal regime shown in the PDF will be the company’s, consistent with the XML.

opw-4989605

closes odoo/enterprise#94817

X-original-commit: 6c604767b55958ad1abde5bb16ed5d39da014b63
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>
Signed-off-by: Guillaume Teboul-Tornezy (gute) <gute@odoo.com>