[IMP] sale_commission*: Prepare to allow materialized achievements
This commits aims to ease the possibility to materialize the
achievements. The achievement table was a report built on the fly.
Several commits were introduced to speed up the performances but on big
systems with thousands of invoices, reading the achievements takes a few
seconds.

On big databases, it may be needed to materialized the achievement view
to speed up reads. This commit reorganize the code to ease the materialization in a simple override commit.

task-5170644

closes odoo/enterprise#97268

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>