[IMP] ai(_fields): add language instruction
Purpose:
--------
Make it less likely that the LLM generates an answer in another language
than the one used in the user's prompt.
Removed translated strings as there is a chance that the user does not
interact with AI in the language set on its user profile (or the technical
user's language in the case of automated actions), in which case we could
have 3 different languages in the prompt.

Task-4989962

Part-of: odoo/enterprise#91842
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>