[FIX] point_of_sale: no prep category selected
Steps to reproduce:
- Open a pos restaurant config that has no prep printer/display.
- Add an orderline.
- The order button appear and if you try to pay the popup ask for send to
preparation is shown.

Issue:
If there is no preparationCategories for a config getOrderChanges consider
that all the available categories are the preparationCategories.

Fix:
If there is no preparationCategories, set the orderline uiState hasChange
to false.

Note:
When no preparation printer category is defined, no categories is to be
return by default. For the preparation display, if no preparation
categories is selected preparationCategories will return all the available
categories.

Task-5016231

closes odoo/enterprise#94672

X-original-commit: 7592034b27714801eab0fe1ba4a99de57b6a9216
Related: odoo/odoo#227025
Signed-off-by: Loukas Wets (lowe) <lowe@odoo.com>