[FIX] account_reports: external value check lock date
This commit add the check that protects external values from being
edited out of the lock dates.
For example when the tax report is submitted, the user is not supposed
to modify any external values anymore.

To do this, we had to modify the tax closing flow a little bit: when
closing the tax period, we now generate the default external values
before setting the tax lock date. This is because the generation of the
default external values was done for the period we were closing, but now
that we forbid the creation of an external value after the lock date we
had to change the order of the flow.

Due to one specific corner case (l10n_fr), we had to keep a hack to
bypass the Tax Return Lock Date check. This was done with a context key
and will have to be removed in master. The case is the following : when
the user generates the tax closing entry, the external values for the
period are generated and the Tax Return Lock Date is set with the last
day of the month. Then if the user tries to submit the EDI VAT report,
it tries to create 2 external values for the carryover but as the lock
date was set, it raises an error.

task-5012442

closes odoo/enterprise#94853

X-original-commit: 81590571c168072fcda83f0c48b213238e0fb2b2
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Romain Toulmonde (roto) <roto@odoo.com>