[IMP] ai: allow quick creation of tools from topics
Purpose:
--------
Make it possible to quick create ai tools from the ai topic form view.
Since only tools with `use_in_ai` can be used in topics, the value is set
by default and is readonly, and some states (that can not be set as usable
in ai) are hidden.

Task-5015090

Part-of: odoo/enterprise#92466
Related: odoo/upgrade#8240
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>