[FIX] timesheet_grid: TZ issue when stopping timer
Before this commit, if you were to log time on a
timesheet using the timer start and stop buttons,
you could potentially log time for the previous
day's or the upcoming day's timesheet entry.
This is because we were searching for
previously-created timesheet entries based on the
UTC Date of the timer's "stop" time.

This fix uses the client's local time to get the
Date of the timer stop. With this, we'll only find
timesheet entries with the `date` field matching
the client's local Date.

closes odoo/enterprise#93549

X-original-commit: 7096753da8fb2bcb6e2a1ed21a2296f7d9617033
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Harrison Hutton (hahu) <hahu@odoo.com>