[IMP] project_enterprise: daterange_picker: date range icon in picker
Before this commit, we had to click on “Add start/end date” next
to the input field.

This has now been replaced with an icon in the date picker,
which you can toggle easily without closing the picker.
It also means that the user no longer needs to know whether it
is a start or an end date.

These tests have been removed because the button is no longer there,
and the icon is always present. This part is already tested in web.

task-4613140

closes odoo/enterprise#93286

Related: odoo/odoo#224439
Signed-off-by: Pierre Paridans (app) <app@odoo.com>