[FIX] account_accountant: fix matching rule for foreign exchange fees
In case of foreign payment using a bank account in another currency, banks can reflect the foreign currency exchange fee on another bank account belonging to the company and in its main currency.
Fr example: in a EUR company having in the same bank (e.g ING) both an USD and a EUR account, making a payment using the USD bank account gets you effectively 2 bank transactions while you initiated only one:
 - the payment in the USD account and
 - the currency diff fee in the EUR account.
In this particular use case, the communication on the currency diff fee reflects the one from the payment and, depending on the reconciliation order, that currency diff could be taken as a partial payment
because we were matching on the words in payment_ref regardless of the currency used, while it shouldn't and must be processed as a bank fee.

Part-of: odoo/enterprise#93025
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>