[IMP] sale_subscription: ux changes for recurring prices
From this PR onwards, when adding a new recurring price on a product,
the usual dialog window with the choice for complex pricing isn't shown
anymore. The list view is used to add new prices and it defaults to
a fixed price.
The prices that are not of the fixed-price type are removed from that
list view since they would show up with a price of 0.

The position of the Pricelist and of the Plan has also been inverted.

closes odoo/enterprise#92613

Task: 5021972
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>